requirements.txt
fastapi
uvicorn
pydantic[dotenv]
python-jose[cryptography]
PyGithub
httpx
python-dotenv
sentence-transformers
torch
faiss-cpu
numpy